Update to timer/goal and some audio assets
Tried to add Events to tie the "Goal" Script, and TimeOut event when counter down to zero. (sorry, haven't used an event system before, tried to plug in code to trigger necessary stuff but not sure)
No success in using HUD to display countdown timer.
Added some audio assets that can be used. All free with CC0 or equivalent license so definitely can be used